Edited Folder Structure
I deleted the assets directory and added app, static, and template to suite
coding with flask. I also moved the styles and script dirs into the static dir.
I also activated venv and run pip freeze to create a requirements file for all the
libraries I will be using for this project.